especificando tipo de retorno da media

+++ media_gols.c
@@ -8,7 +8,7 @@ int main() {
     scanf("%d%*c", &qtdGols);
     printf("Digite a quantidade de partidas jogadas: ");
     scanf("%d%*c", &qtdPartidas);
-    mediaGols = qtdGols / qtdPartidas;
+    mediaGols = (float)qtdGols / qtdPartidas;
     printf("\nMedia de gols por partida = %.0f\n", mediaGols);
 
     return 0;
